보호된 경로에 대한 인증 미들웨어 구현
